Scaffold new version
Scaffold the new version of the site.
Remove components and routes that will be rewritten.
Remove `react-router`, `react-icons` and `sass`.
Add `mantine` and `tabler-icons` for easy components. Switch to CSS
modules over SCSS because of mantine.
Add linting and some base rules.
Add prettier.
Add a basic layout with home and projects pages.